Refactoring ACC CUD parts.

+++ src/main/java/org/oagi/srt/gateway/http/api/cc_management/data/CcAccCreateRequest.java
@@ -2,9 +2,11 @@ package org.oagi.srt.gateway.http.api.cc_management.data;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class CcAccCreateRequest {
 
-    private long releaseId;
+    private BigInteger releaseId;
 
 }
